api/requirements.txt
aniso8601==9.0.1
cffi==1.15.0
click==8.0.3
cryptography==35.0.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
uWSGI==2.0.20
Werkzeug==2.0.2
